Prevent overflow in highbd_filter_intra_predictors_*tap()
BUG=aomedia:418

+++ av1/common/reconintra.c
@@ -1560,16 +1560,16 @@ static void highbd_filter_intra_predictors_3tap(uint16_t *dst, ptrdiff_t stride,
                                                 const uint16_t *left, int mode,
                                                 int bd) {
   int k, r, c;
-  int mean, ipred;
+  int64_t mean, ipred;
 #if CONFIG_TX64X64
-  int preds[65][65];
+  int64_t preds[65][65];
 #else
-  int preds[33][33];
+  int64_t preds[33][33];
 #endif  // CONFIG_TX64X64
   const TX_SIZE tx_size = get_txsize_from_blocklen(bs);
-  const int c0 = av1_filter_intra_taps_3[tx_size][mode][0];
-  const int c1 = av1_filter_intra_taps_3[tx_size][mode][1];
-  const int c2 = av1_filter_intra_taps_3[tx_size][mode][2];
+  const int64_t c0 = av1_filter_intra_taps_3[tx_size][mode][0];
+  const int64_t c1 = av1_filter_intra_taps_3[tx_size][mode][1];
+  const int64_t c2 = av1_filter_intra_taps_3[tx_size][mode][2];
 
   k = 0;
   mean = 0;
@@ -1605,17 +1605,17 @@ static void highbd_filter_intra_predictors_4tap(uint16_t *dst, ptrdiff_t stride,
                                                 const uint16_t *left, int mode,
                                                 int bd) {
   int k, r, c;
-  int mean, ipred;
+  int64_t mean, ipred;
 #if CONFIG_TX64X64
-  int preds[65][129];
+  int64_t preds[65][129];
 #else
-  int preds[33][65];
+  int64_t preds[33][65];
 #endif  // CONFIG_TX64X64
   const TX_SIZE tx_size = get_txsize_from_blocklen(bs);
-  const int c0 = av1_filter_intra_taps_4[tx_size][mode][0];
-  const int c1 = av1_filter_intra_taps_4[tx_size][mode][1];
-  const int c2 = av1_filter_intra_taps_4[tx_size][mode][2];
-  const int c3 = av1_filter_intra_taps_4[tx_size][mode][3];
+  const int64_t c0 = av1_filter_intra_taps_4[tx_size][mode][0];
+  const int64_t c1 = av1_filter_intra_taps_4[tx_size][mode][1];
+  const int64_t c2 = av1_filter_intra_taps_4[tx_size][mode][2];
+  const int64_t c3 = av1_filter_intra_taps_4[tx_size][mode][3];
 
   k = 0;
   mean = 0;
